Add new crate databend-common-meta-app-types, defining the types used by meta-service application, that depend on meta-service but meta-service does not depend on them.
move NonEmptyString to app-types crate.

remove unused DATABEND_COMMIT_VERSION from meta-types crate.

Extract `Node` from mod `cluster` to mod `node`.
