+++ Cargo.toml
@@ -91,6 +91,7 @@ members = [
     "src/query/ee",
     "src/meta/api",
     "src/meta/app",
+    "src/meta/app-types",
     "src/meta/binaries",
     "src/meta/client",
     "src/meta/control",
@@ -138,6 +139,7 @@ databend-common-license = { path = "src/common/license" }
 databend-common-management = { path = "src/query/management" }
 databend-common-meta-api = { path = "src/meta/api" }
 databend-common-meta-app = { path = "src/meta/app" }
+databend-common-meta-app-types = { path = "src/meta/app-types" }
 databend-common-meta-client = { path = "src/meta/client" }
 databend-common-meta-control = { path = "src/meta/control" }
 databend-common-meta-embedded = { path = "src/meta/embedded" }
