+++ src/meta/app-types/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "databend-common-meta-app-types"
+version = { workspace = true }
+authors = { workspace = true }
+license = { workspace = true }
+publish = { workspace = true }
+edition = { workspace = true }
+
+[lib]
+doctest = false
+test = true
+
+[dependencies]
+num-derive = { workspace = true }
+prost = { workspace = true }
+serde = { workspace = true }
+
+[dev-dependencies]
+anyhow = { workspace = true }
+
+[package.metadata.cargo-machete]
+ignored = ["num-derive", "prost"]
+
+[lints]
+workspace = true
