+++ src/meta/app-types/README.md
@@ -1,0 +1,2 @@
+Defines types used by meta-service applications.
+In other words, these types depend on the meta-service but meta-service does not depend on them.
