+++ src/meta/app-types/src/lib.rs
@@ -12,16 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::sync::LazyLock;
-
-pub static DATABEND_COMMIT_VERSION: LazyLock<String> = LazyLock::new(|| {
-    let build_semver = option_env!("DATABEND_GIT_SEMVER");
-    let git_sha = option_env!("VERGEN_GIT_SHA");
-    let rustc_semver = option_env!("VERGEN_RUSTC_SEMVER");
-    let timestamp = option_env!("VERGEN_BUILD_TIMESTAMP");
-
-    match (build_semver, git_sha, rustc_semver, timestamp) {
-        (Some(v1), Some(v2), Some(v3), Some(v4)) => format!("{}-{}({}-{})", v1, v2, v3, v4),
-        _ => String::new(),
-    }
-});
+pub mod non_empty;
