+++ src/meta/app-types/src/non_empty.rs
