+++ src/meta/app/src/tenant/uninit_tenant.rs
@@ -14,7 +14,7 @@
 
 use std::fmt::Display;
 
-use databend_common_meta_types::NonEmptyString;
+use databend_common_meta_app_types::non_empty::NonEmptyString;
 
 use crate::app_error::TenantIsEmpty;
 use crate::tenant::Tenant;
