+++ src/meta/app/src/tenant_key/ident.rs
@@ -186,10 +186,10 @@ where
 mod kvapi_key_impl {
     use std::fmt::Debug;
 
+    use databend_common_meta_app_types::non_empty::NonEmptyString;
     use databend_common_meta_kvapi::kvapi;
     use databend_common_meta_kvapi::kvapi::KeyCodec;
     use databend_common_meta_kvapi::kvapi::KeyError;
-    use databend_common_meta_types::NonEmptyString;
 
     use crate::tenant::Tenant;
     use crate::tenant_key::ident::TIdent;
