+++ src/meta/app/tests/it/main.rs
@@ -17,10 +17,3 @@ mod user_grant;
 mod user_info;
 mod user_privilege;
 mod user_quota;
-
-#[test]
-fn test_bin_commit_version() -> anyhow::Result<()> {
-    let v = &databend_common_meta_types::config::DATABEND_COMMIT_VERSION;
-    assert!(v.len() > 0);
-    Ok(())
-}
