+++ src/meta/binaries/meta/entry.rs
@@ -26,10 +26,10 @@ use databend_common_meta_raft_store::ondisk::OnDisk;
 use databend_common_meta_raft_store::ondisk::DATA_VERSION;
 use databend_common_meta_sled_store::openraft::MessageSummary;
 use databend_common_meta_store::MetaStoreProvider;
+use databend_common_meta_types::node::Node;
 use databend_common_meta_types::Cmd;
 use databend_common_meta_types::LogEntry;
 use databend_common_meta_types::MetaAPIError;
-use databend_common_meta_types::Node;
 use databend_common_tracing::init_logging;
 use databend_common_tracing::set_panic_hook;
 use databend_meta::api::GrpcServer;
