+++ src/meta/control/src/import.rs
@@ -37,6 +37,7 @@ use databend_common_meta_raft_store::state_machine::MetaSnapshotId;
 use databend_common_meta_sled_store::init_get_sled_db;
 use databend_common_meta_sled_store::openraft::storage::RaftLogStorageExt;
 use databend_common_meta_sled_store::openraft::RaftSnapshotBuilder;
+use databend_common_meta_types::node::Node;
 use databend_common_meta_types::raft_types::CommittedLeaderId;
 use databend_common_meta_types::raft_types::Entry;
 use databend_common_meta_types::raft_types::EntryPayload;
@@ -48,7 +49,6 @@ use databend_common_meta_types::sys_data::SysData;
 use databend_common_meta_types::Cmd;
 use databend_common_meta_types::Endpoint;
 use databend_common_meta_types::LogEntry;
-use databend_common_meta_types::Node;
 use databend_meta::store::RaftStore;
 use raft_log::api::raft_log_writer::RaftLogWriter;
 use url::Url;
