+++ src/meta/kvapi/Cargo.toml
@@ -16,6 +16,7 @@ test = true
 anyhow = { workspace = true }
 async-trait = { workspace = true }
 databend-common-base = { workspace = true }
+databend-common-meta-app-types = { workspace = true }
 databend-common-meta-types = { workspace = true }
 fastrace = { workspace = true }
 futures-util = { workspace = true }
