+++ src/meta/proto-conv/Cargo.toml
@@ -16,6 +16,7 @@ chrono-tz = { workspace = true, features = ["serde"] }
 databend-common-expression = { workspace = true }
 databend-common-io = { workspace = true }
 databend-common-meta-app = { workspace = true }
+databend-common-meta-app-types = { workspace = true }
 databend-common-meta-types = { workspace = true }
 databend-common-protos = { workspace = true }
 enumflags2 = { workspace = true }
