+++ src/meta/proto-conv/src/lock_from_to_protobuf_impl.rs
@@ -19,7 +19,7 @@ use chrono::DateTime;
 use chrono::Utc;
 use databend_common_meta_app::schema as mt;
 use databend_common_meta_app::tenant::Tenant;
-use databend_common_meta_types::NonEmptyString;
+use databend_common_meta_app_types::non_empty::NonEmptyString;
 use databend_common_protos::pb;
 use num::FromPrimitive;
