+++ src/meta/raft-store/src/sm_v003/compact_with_db_test.rs
@@ -14,11 +14,11 @@
 
 use std::io;
 
+use databend_common_meta_types::node::Node;
 use databend_common_meta_types::raft_types::Membership;
 use databend_common_meta_types::raft_types::StoredMembership;
 use databend_common_meta_types::seq_value::KVMeta;
 use databend_common_meta_types::Endpoint;
-use databend_common_meta_types::Node;
 use databend_common_meta_types::UpsertKV;
 use futures_util::TryStreamExt;
 use map_api::map_api::MapApi;
