+++ src/meta/service/src/meta_service/meta_leader.rs
@@ -21,6 +21,7 @@ use databend_common_meta_kvapi::kvapi::KVApi;
 use databend_common_meta_raft_store::sm_v003::SMV003;
 use databend_common_meta_sled_store::openraft::ChangeMembers;
 use databend_common_meta_stoerr::MetaStorageError;
+use databend_common_meta_types::node::Node;
 use databend_common_meta_types::protobuf::StreamItem;
 use databend_common_meta_types::raft_types::ClientWriteError;
 use databend_common_meta_types::raft_types::MembershipNode;
@@ -33,7 +34,6 @@ use databend_common_meta_types::LogEntry;
 use databend_common_meta_types::MetaDataError;
 use databend_common_meta_types::MetaDataReadError;
 use databend_common_meta_types::MetaOperationError;
-use databend_common_meta_types::Node;
 use databend_common_metrics::count::Count;
 use futures::StreamExt;
 use futures::TryStreamExt;
