+++ src/meta/service/src/meta_service/meta_node.rs
@@ -37,6 +37,7 @@ use databend_common_meta_sled_store::openraft;
 use databend_common_meta_sled_store::openraft::error::RaftError;
 use databend_common_meta_sled_store::openraft::ChangeMembers;
 use databend_common_meta_stoerr::MetaStorageError;
+use databend_common_meta_types::node::Node;
 use databend_common_meta_types::protobuf::raft_service_client::RaftServiceClient;
 use databend_common_meta_types::protobuf::raft_service_server::RaftServiceServer;
 use databend_common_meta_types::protobuf::watch_request::FilterType;
@@ -62,7 +63,6 @@ use databend_common_meta_types::MetaManagementError;
 use databend_common_meta_types::MetaNetworkError;
 use databend_common_meta_types::MetaOperationError;
 use databend_common_meta_types::MetaStartupError;
-use databend_common_meta_types::Node;
 use fastrace::func_name;
 use fastrace::prelude::*;
 use itertools::Itertools;
