+++ src/meta/service/src/meta_service/meta_node_status.rs
@@ -16,9 +16,9 @@ use std::collections::BTreeMap;
 
 use databend_common_meta_raft_store::ondisk::DataVersion;
 use databend_common_meta_raft_store::raft_log_v004::RaftLogStat;
+use databend_common_meta_types::node::Node;
 use databend_common_meta_types::raft_types::LogId;
 use databend_common_meta_types::raft_types::NodeId;
-use databend_common_meta_types::Node;
 
 #[derive(serde::Serialize)]
 pub struct MetaNodeStatus {
