+++ src/meta/service/tests/it/api/http/cluster_state_test.rs
@@ -20,7 +20,7 @@ use std::time::Duration;
 use databend_common_base::base::tokio;
 use databend_common_base::base::tokio::time::Instant;
 use databend_common_base::base::Stoppable;
-use databend_common_meta_types::Node;
+use databend_common_meta_types::node::Node;
 use databend_meta::api::http::v1::cluster_state::nodes_handler;
 use databend_meta::api::http::v1::cluster_state::status_handler;
 use databend_meta::api::HttpService;
