+++ src/meta/service/tests/it/tests/meta_node.rs
@@ -19,9 +19,9 @@ use std::sync::Arc;
 use std::time::Duration;
 
 use databend_common_meta_sled_store::openraft::ServerState;
+use databend_common_meta_types::node::Node;
 use databend_common_meta_types::raft_types::NodeId;
 use databend_common_meta_types::AppliedState;
-use databend_common_meta_types::Node;
 use databend_meta::meta_service::MetaNode;
 use log::info;
 use maplit::btreeset;
