+++ src/meta/sled-store/src/sled_serde_impl.rs
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use databend_common_meta_types::node::Node;
 use databend_common_meta_types::raft_types::Entry;
 use databend_common_meta_types::raft_types::EntryPayload;
 use databend_common_meta_types::raft_types::LogId;
@@ -20,7 +21,6 @@ use databend_common_meta_types::raft_types::SnapshotMeta;
 use databend_common_meta_types::raft_types::StoredMembership;
 use databend_common_meta_types::raft_types::Vote;
 use databend_common_meta_types::seq_value::SeqV;
-use databend_common_meta_types::Node;
 use databend_common_meta_types::SeqNum;
 
 use crate::SledBytesError;
