+++ src/meta/types/src/cluster.rs
@@ -12,63 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::fmt;
 use std::net::AddrParseError;
 use std::net::SocketAddr;
 use std::str::FromStr;
 
 use serde::Deserialize;
 use serde::Serialize;
-
-use crate::Endpoint;
-
-#[derive(Serialize, Deserialize, Debug, Default, Clone, PartialEq, Eq, deepsize::DeepSizeOf)]
-pub struct Node {
-    /// Node name for display.
-    pub name: String,
-
-    /// Raft service endpoint to connect to.
-    pub endpoint: Endpoint,
-
-    /// For backward compatibility, it can not be removed.
-    /// 2023-02-09
-    //#[deprecated(note = "it is listening addr, not advertise addr")]
-    #[serde(skip)]
-    pub grpc_api_addr: Option<String>,
-
-    /// The address `ip:port` for a meta-client to connect to.
-    pub grpc_api_advertise_address: Option<String>,
-}
-
-impl Node {
-    pub fn new(name: impl ToString, endpoint: Endpoint) -> Self {
-        Self {
-            name: name.to_string(),
-            endpoint,
-            ..Default::default()
-        }
-    }
-
-    pub fn with_grpc_advertise_address(mut self, g: Option<impl ToString>) -> Self {
-        self.grpc_api_advertise_address = g.map(|x| x.to_string());
-        self
-    }
-}
-
-impl fmt::Display for Node {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let grpc_addr_display = if let Some(grpc_addr) = &self.grpc_api_advertise_address {
-            grpc_addr.to_string()
-        } else {
-            "".to_string()
-        };
-        write!(
-            f,
-            "id={} raft={} grpc={}",
-            self.name, self.endpoint, grpc_addr_display
-        )
-    }
-}
 
 #[derive(Serialize, Deserialize, Clone, Debug, Eq, PartialEq, Ord, PartialOrd, Default)]
 pub enum NodeType {
@@ -180,7 +129,7 @@ impl NodeInfo {
 
 #[cfg(test)]
 mod tests {
-    use crate::Node;
+    use crate::node::Node;
 
     #[test]
     fn test_serde_node_compatible() -> anyhow::Result<()> {
