+++ src/meta/types/src/lib.rs
@@ -27,7 +27,6 @@ mod grpc_helper;
 mod log_entry;
 mod match_seq;
 mod message;
-mod non_empty;
 mod operation;
 mod raft_snapshot_data;
 mod raft_txid;
@@ -40,13 +39,12 @@ mod proto_display;
 mod proto_ext;
 
 pub mod cmd;
-pub mod config;
 pub mod errors;
+pub mod node;
 pub mod raft_types;
 pub mod seq_value;
 pub mod snapshot_db;
 pub mod sys_data;
-
 // reexport
 
 pub use anyerror;
@@ -61,7 +59,6 @@ pub mod protobuf {
 
 pub use applied_state::AppliedState;
 pub use change::Change;
-pub use cluster::Node;
 pub use cluster::NodeInfo;
 pub use cluster::NodeType;
 pub use endpoint::Endpoint;
@@ -85,6 +82,7 @@ pub use log_entry::LogEntry;
 pub use map_api::expirable::Expirable;
 pub use match_seq::MatchSeq;
 pub use match_seq::MatchSeqExt;
+pub use node::Node;
 pub use operation::MetaId;
 pub use operation::Operation;
 pub use proto_display::VecDisplay;
@@ -120,6 +118,4 @@ pub use crate::cmd::CmdContext;
 pub use crate::cmd::MetaSpec;
 pub use crate::cmd::UpsertKV;
 pub use crate::grpc_helper::GrpcHelper;
-pub use crate::non_empty::NonEmptyStr;
-pub use crate::non_empty::NonEmptyString;
 pub use crate::raft_snapshot_data::SnapshotData;
