+++ src/meta/types/tests/it/main.rs
@@ -14,10 +14,3 @@
 
 mod cluster;
 mod txn_serde;
-
-#[test]
-fn test_bin_commit_version() -> anyhow::Result<()> {
-    let v = &databend_common_meta_types::config::DATABEND_COMMIT_VERSION;
-    assert!(v.len() > 0);
-    Ok(())
-}
